Completed lemma 4
Please check that when we say "given a rootpairing, the map s is unique", whether we mean that if P and Q are both rootpairings with the same R M N L I root coroot, then they have the same s. So I made the three assumptions h1 h2 h3. I wonder if there's any cleaner way to state it.